+++ corehq/apps/app_manager/decorators.py
@@ -15,10 +15,6 @@ from couchdbkit.exceptions import ResourceConflict
 from dimagi.utils.couch.undo import DELETED_SUFFIX
 
 from corehq import toggles
-from corehq.apps.app_manager.const import (
-    PUBLIC_FORM_SESSION_COOKIE_NAME,
-    PUBLIC_FORM_SESSION_HEADER,
-)
 from corehq.apps.app_manager.dbaccessors import get_app
 from corehq.apps.app_manager.exceptions import (
     AppEditingError,
@@ -28,7 +24,6 @@ from corehq.apps.app_manager.exceptions import (
 from corehq.apps.app_manager.util import get_latest_app_release_by_location
 from corehq.apps.domain.decorators import login_and_domain_required
 from corehq.apps.domain.models import Domain
-from corehq.apps.public_webforms.models import PublicFormUser, PublicFormSession
 from corehq.apps.users.decorators import require_permission
 from corehq.apps.users.models import CommCareUser, HqPermissions
 from corehq.apps.users.util import normalize_username
@@ -137,33 +132,6 @@ def safe_cached_download(f):
         messages.error(request, "Problem downloading file: %s" % message)
         return HttpResponseRedirect(reverse("view_app", args=[domain, app_id]))
     return _safe_cached_download
-
-
-def allow_public_form_session(view_func):
-    """
-    Adds public form session auth as an accepted auth mode for a view.
-
-    When the request carries the public form session header and a cookie
-    whose key resolves to a valid, usable PublicFormSession, sets
-    ``request.couch_user`` to a PublicFormUser proxy for that session.
-    """
-    @wraps(view_func)
-    def _inner(request, *args, **kwargs):
-        session = _get_public_form_session(request)
-        if session is not None:
-            request.couch_user = PublicFormUser(session)
-        return view_func(request, *args, **kwargs)
-
-    return _inner
-
-
-def _get_public_form_session(request):
-    if request.headers.get(PUBLIC_FORM_SESSION_HEADER) != 'true':
-        return None
-    raw_key = request.COOKIES.get(PUBLIC_FORM_SESSION_COOKIE_NAME)
-    if not raw_key:
-        return None
-    return PublicFormSession.get_active_session_by_key(raw_key)
 
 
 def no_conflict_require_POST(fn):
