+++ corehq/apps/app_manager/tests/test_public_webforms.py
@@ -11,7 +11,7 @@ from corehq.apps.app_manager.const import (
     PUBLIC_FORM_SESSION_COOKIE_NAME,
     PUBLIC_FORM_SESSION_HEADER,
 )
-from corehq.apps.app_manager.decorators import allow_public_form_session
+from corehq.apps.public_webforms.decorators import allow_public_form_session
 from corehq.apps.public_webforms.models import (
     OTARestorePublicFormUser,
     PublicFormSession,
